+++ app/models/implementing_organisation.rb
@@ -1,0 +1,5 @@
+class ImplementingOrganisation < ApplicationRecord
+  validates_presence_of :name, :organisation_type
+
+  belongs_to :activity
+end