+++ spec/models/activity_spec.rb
@@ -133,6 +133,7 @@ RSpec.describe Activity, type: :model do
     it { should belong_to(:activity).optional }
     it { should have_many(:activities).with_foreign_key("activity_id") }
     it { should belong_to(:extending_organisation).with_foreign_key("extending_organisation_id").optional }
+    it { should have_many(:implementing_organisations) }
   end
 
   describe "#parent_activity" do
@@ -206,5 +207,13 @@ RSpec.describe Activity, type: :model do
     activity = build(:activity)
 
     expect(activity.has_extending_organisation?).to be false
+  end
+
+  describe "#has_implementing_organisation?" do
+    it "returns true when there is one or more implementing organisationg" do
+      activity = create(:project_activity_with_implementing_organisations)
+
+      expect(activity.has_implementing_organisations?).to be true
+    end
   end
 end